Dont change dtypes. Its bad for your memory

+++ gpaw/core/plane_waves.py
@@ -352,7 +352,8 @@ class PWArray(DistributedArrays[PWDesc]):
         return PWArray(self.desc,
                        self.dims,
                        self.comm,
-                       data)
+                       data,
+                       dtype=self.dtype)
 
     def copy(self):
         """Create a copy (surprise!)."""
